refactor: split landing page into server + client components
Extract static landing page content (hero, demo cards, features, funnel,
CTA) into LandingContent Server Component. The page.tsx wrapper now only
handles client-side concerns: auth state check, OAuth redirect, and GSAP
animations. Reduces client JS bundle for the landing page.
